requirements.txt
azureml-sdk[notebooks]
azureml-mlflow
pytorch-forecasting
pytorch-lightning
scikit-learn
pandas
matplotlib
seaborn
mlflow
peft
transformers
bitsandbytes
chroma
openai
langchain
streamlit
fastapi
pytest
requests
jupyter
ipykernel
